backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "cachetools>=6.1.0",
    "fastapi[standard]>=0.116.1",
    "gradio>=5.42.0",
    "haystack-ai>=2.16.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=2.3.2",
    "ollama-haystack>=4.1.0",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
